Fixed account registration bugs
Reworked the entire registration page backend to make it more user
friendly.
Fixed typo in reCAPTCHA text.  Should be working now as intended.
Fixed several SQL query functions that needed row counts.
Fixed the built in redirect function to correctly accommodate meta
redirects.
Added an option to disable the "email sent" notice on the webpage when
an email was sent.
Issue fixed #24 where SQL error was displayed when a user tried to
register an existing username.